I2C Protocol refactor v2
Updated I2C protocol to operate on single commands instead of tokens.
* Driver design now generic (in i2c_driver.h) implementing a finite
  state machine.
* Virtualiser now handles multicore and ensures fairness as this is
  necessary for functional I2C.
* Added libi2c, a user-level library for easy blocking I2C calls.
* Significant cleanup and general improvements in code quality.

+++ include/sddf/i2c/devices/pn532/pn532.h
@@ -29,5 +29,5 @@
 
 #define PN532_MIFARE_ISO14443A          (0x00)
 
-uint8_t pn532_write_command(uint8_t *header, uint8_t hlen, const uint8_t *body, uint8_t blen, size_t retries);
-uint8_t pn532_read_response(uint8_t *buffer, uint8_t buffer_len, size_t retries);
+uint8_t pn532_write_command(uint8_t *header, uint8_t hlen, const uint8_t *body, uint8_t blen, unsigned retries);
+uint8_t pn532_read_response(uint8_t *buffer, uint8_t buffer_len, unsigned retries);
